requirements.txt
streamlit
pillow
ultralytics